chore: refactored application submission handlers, services and repository
