+++ apps/api/internal/api/handlers/events.go
@@ -2,19 +2,15 @@ package handlers
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 	"reflect"
-	"strconv"
 	"time"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-playground/validator/v10"
 	"github.com/google/uuid"
 	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
-	"github.com/swamphacks/core/apps/api/internal/api/middleware"
 	res "github.com/swamphacks/core/apps/api/internal/api/response"
 	"github.com/swamphacks/core/apps/api/internal/config"
 	"github.com/swamphacks/core/apps/api/internal/ctxutils"
@@ -391,256 +387,6 @@ func (h *EventHandler) AssignEventRole(w http.ResponseWriter, r *http.Request) {
 		}
 		return
 	}
-
-	w.WriteHeader(http.StatusOK)
-}
-
-func getEventID(w http.ResponseWriter, r *http.Request) (uuid.UUID, error) {
-	eventIdStr := chi.URLParam(r, "eventId")
-	if eventIdStr == "" {
-		err := errors.New("missing_event_id")
-		errMsg := "The event ID is missing from the URL!"
-
-		log.Err(err).Msg(errMsg)
-
-		res.SendError(w, http.StatusBadRequest,
-			res.NewError(err.Error(), errMsg))
-
-		return uuid.Nil, err
-	}
-
-	eventId, err := uuid.Parse(eventIdStr)
-	if err != nil {
-		err = errors.New("invalid_event_id")
-		errMsg := "The event ID is not a valid UUID"
-
-		log.Err(err).Msg(errMsg)
-
-		res.SendError(w, http.StatusBadRequest,
-			res.NewError(err.Error(), errMsg))
-
-		return uuid.Nil, err
-	}
-
-	return eventId, nil
-}
-
-func getUserContext(w http.ResponseWriter, r *http.Request) (*middleware.UserContext, error) {
-	userContext, ok := r.Context().Value(middleware.UserContextKey).(*middleware.UserContext)
-
-	if !ok || userContext == nil {
-		err := errors.New("user_error")
-
-		log.Err(err).Msg("unable to fetch user info")
-
-		res.SendError(w, http.StatusBadRequest, res.NewError(err.Error(), "Unable to fetch user information."))
-		return nil, err
-	}
-
-	return userContext, nil
-}
-
-func (h *EventHandler) GetApplicationByUserAndEventID(w http.ResponseWriter, r *http.Request) {
-	eventId, err := getEventID(w, r)
-
-	if err != nil {
-		return
-	}
-
-	userContext, err := getUserContext(w, r)
-
-	if err != nil {
-		return
-	}
-
-	params := sqlc.GetApplicationByUserAndEventIDParams{
-		UserID:  userContext.UserID,
-		EventID: eventId,
-	}
-
-	application, err := h.eventService.GetApplicationByUserAndEventID(r.Context(), params)
-
-	if err != nil {
-		if err == repository.ErrApplicationNotFound {
-			params := sqlc.CreateApplicationParams{
-				UserID:  userContext.UserID,
-				EventID: eventId,
-			}
-			newApplication, err := h.eventService.CreateApplication(r.Context(), params)
-			if err != nil {
-				res.SendError(w, http.StatusBadRequest, res.NewError("create_application_error", "can't create application"))
-			} else {
-				res.Send(w, http.StatusOK, newApplication)
-			}
-			res.SendError(w, http.StatusBadRequest, res.NewError("application_not_found", "can't find application"))
-		} else {
-			res.SendError(w, http.StatusBadRequest, res.NewError("get_application_error", "error retrieving application"))
-		}
-
-		return
-	}
-
-	res.Send(w, http.StatusOK, application)
-}
-
-// TODO: figure out a way to create the submission fields dynamically using the json form files with proper validation.
-// these fields are only applicable to swamphacks xi, not other events
-type ApplicationSubmissionFields struct {
-	FirstName               string `json:"firstName" validate:"required,max=50"`
-	LastName                string `json:"lastName" validate:"required,max=50"`
-	Age                     int    `json:"age" validate:"required,min=0,max=99"`
-	Phone                   string `json:"phone" validate:"required,len=10"`
-	PreferredEmail          string `json:"preferredEmail" validate:"required,email"`
-	UniversityEmail         string `json:"universityEmail" validate:"required,email"`
-	Linkedin                string `json:"linkedin" validate:"required,url"`
-	Github                  string `json:"github" validate:"required,url"`
-	AgeCertification        bool   `json:"ageCertification" validate:"required,boolean"`
-	School                  string `json:"school" validate:"required"`
-	Level                   string `json:"level" validate:"required"`
-	Year                    string `json:"year" validate:"required"`
-	GraduationYear          string `json:"graduationYear" validate:"required"`
-	Majors                  string `json:"majors" validate:"required"`
-	Minors                  string `json:"minors"`
-	Experience              string `json:"experience" validate:"required"`
-	ProjectExperience       string `json:"projectExperience" validate:"required"`
-	ShirtSize               string `json:"shirtSize" validate:"required"`
-	Essay1                  string `json:"essay1" validate:"required"`
-	Essay2                  string `json:"essay2" validate:"required"`
-	Referral                string `json:"referral" validate:"required"`
-	PictureConsent          string `json:"pictureConsent" validate:"required"`
-	InpersonAcknowledgement string `json:"inpersonAcknowledgement" validate:"required"`
-	AgreeToConduct          string `json:"agreeToConduct" validate:"required"`
-	InfoShareAuthorization  string `json:"infoShareAuthorization" validate:"required"`
-	AgreeToMLHEmails        string `json:"agreeToMLHEmails" validate:"required"`
-}
-
-func (h *EventHandler) SubmitApplication(w http.ResponseWriter, r *http.Request) {
-	// Parse multipart form (10 MB max memory)
-	err := r.ParseMultipartForm(10 << 20)
-	if err != nil {
-		http.Error(w, "Failed to parse form: "+err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	var submission ApplicationSubmissionFields
-
-	// Map form values
-	submission.FirstName = r.FormValue("firstName")
-	submission.LastName = r.FormValue("lastName")
-
-	if ageStr := r.FormValue("age"); ageStr != "" {
-		if age, err := strconv.Atoi(ageStr); err == nil {
-			submission.Age = age
-		}
-	}
-
-	submission.Phone = r.FormValue("phone")
-	submission.PreferredEmail = r.FormValue("preferredEmail")
-	submission.UniversityEmail = r.FormValue("universityEmail")
-	submission.Linkedin = r.FormValue("linkedin")
-	submission.Github = r.FormValue("github")
-
-	if ageCertStr := r.FormValue("ageCertification"); ageCertStr != "" {
-		submission.AgeCertification = (ageCertStr == "true" || ageCertStr == "1")
-	}
-
-	submission.School = r.FormValue("school")
-	submission.Level = r.FormValue("level")
-	submission.Year = r.FormValue("year")
-	submission.GraduationYear = r.FormValue("graduationYear")
-	submission.Majors = r.FormValue("majors")
-	submission.Minors = r.FormValue("minors")
-	submission.Experience = r.FormValue("experience")
-	submission.ProjectExperience = r.FormValue("projectExperience")
-	submission.ShirtSize = r.FormValue("shirtSize")
-	submission.Essay1 = r.FormValue("essay1")
-	submission.Essay2 = r.FormValue("essay2")
-	submission.Referral = r.FormValue("referral")
-	submission.PictureConsent = r.FormValue("pictureConsent")
-	submission.InpersonAcknowledgement = r.FormValue("inpersonAcknowledgement")
-	submission.AgreeToConduct = r.FormValue("agreeToConduct")
-	submission.InfoShareAuthorization = r.FormValue("infoShareAuthorization")
-	submission.AgreeToMLHEmails = r.FormValue("agreeToMLHEmails")
-
-	// TODO: Handle file upload to r2
-
-	// file, header, err := r.FormFile("resume[]")
-	// if err == nil {
-	// 	defer file.Close()
-
-	// 	// // Ensure directory exists
-	// 	if err := os.MkdirAll("./uploads", 0755); err != nil {
-	// 		http.Error(w, err.Error(), http.StatusInternalServerError)
-	// 		return
-	// 	}
-
-	// 	// Example: save to local disk (you might upload to S3/GCS instead)
-	// 	dst, err := os.Create("./uploads/" + header.Filename)
-	// 	if err != nil {
-	// 		http.Error(w, "Failed to save resume: "+err.Error(), http.StatusInternalServerError)
-	// 		return
-	// 	}
-	// 	defer dst.Close()
-
-	// 	if _, err := io.Copy(dst, file); err != nil {
-	// 		http.Error(w, "Failed to copy resume: "+err.Error(), http.StatusInternalServerError)
-	// 		return
-	// 	}
-	// } else if err != http.ErrMissingFile {
-	// 	http.Error(w, "Error reading resume: "+err.Error(), http.StatusInternalServerError)
-	// 	return
-
-	// }
-
-	validate := validator.New()
-	if err := validate.Struct(submission); err != nil {
-		res.SendError(w, http.StatusBadRequest, res.NewError("invalid_request", err.Error()))
-		return
-	}
-
-	eventId, err := getEventID(w, r)
-
-	if err != nil {
-		return
-	}
-
-	userContext, err := getUserContext(w, r)
-
-	if err != nil {
-		return
-	}
-
-	userId := userContext.UserID
-	h.eventService.SubmitApplication(r.Context(), submission, userId, eventId)
-
-	w.WriteHeader(http.StatusOK)
-}
-
-func (h *EventHandler) SaveApplication(w http.ResponseWriter, r *http.Request) {
-	var data any
-
-	if err := json.NewDecoder(r.Body).Decode(&data); err != nil {
-		res.SendError(w, http.StatusBadRequest, res.NewError("invalid_form_data", "Something went wrong while parsing form submission"))
-		return
-	}
-
-	eventId, err := getEventID(w, r)
-
-	if err != nil {
-		return
-	}
-
-	userContext, err := getUserContext(w, r)
-
-	if err != nil {
-		return
-	}
-
-	userId := userContext.UserID
-	h.eventService.SaveApplication(r.Context(), data, sqlc.UpdateApplicationParams{
-		UserID:  userId,
-		EventID: eventId,
-	})
 
 	w.WriteHeader(http.StatusOK)
 }
