+++ apps/api/internal/db/repository/events.go
@@ -3,7 +3,6 @@ package repository
 import (
 	"context"
 	"database/sql"
-	"encoding/json"
 	"errors"
 
 	"github.com/google/uuid"
@@ -13,18 +12,12 @@ import (
 )
 
 var (
-	ErrEventNotFound          = errors.New("event not found")
-	ErrEventRoleNotFound      = errors.New("event role not found")
-	ErrDuplicateEvent         = errors.New("event already exists in database")
-	ErrNoEventsDeleted        = errors.New("no events deleted")
-	ErrMultipleEventsDeleted  = errors.New("multiple events affected by delete query while only expecting one to delete one")
-	ErrCreateApplication      = errors.New("unable to create application")
-	ErrSaveApplication        = errors.New("unable to save application")
-	ErrSubmitApplication      = errors.New("unable to submit application")
-	ErrInvalidApplicationData = errors.New("unable to parse application data")
-	ErrGetApplication         = errors.New("unable to get application for user")
-	ErrApplicationNotFound    = errors.New("can not find application for user")
-	ErrUnknown                = errors.New("an unkown error was caught")
+	ErrEventNotFound         = errors.New("event not found")
+	ErrEventRoleNotFound     = errors.New("event role not found")
+	ErrDuplicateEvent        = errors.New("event already exists in database")
+	ErrNoEventsDeleted       = errors.New("no events deleted")
+	ErrMultipleEventsDeleted = errors.New("multiple events affected by delete query while only expecting one to delete one")
+	ErrUnknown               = errors.New("an unkown error was caught")
 )
 
 type EventRepository struct {
@@ -123,82 +116,4 @@ func (r *EventRepository) GetEventStaff(ctx context.Context, eventId uuid.UUID) 
 
 func (r *EventRepository) AssignRole(ctx context.Context, params sqlc.AssignRoleParams) error {
 	return r.db.Query.AssignRole(ctx, params)
-}
-
-func (r *EventRepository) CreateApplication(ctx context.Context, params sqlc.CreateApplicationParams) (*sqlc.Application, error) {
-	application, err := r.db.Query.CreateApplication(ctx, params)
-
-	if err != nil {
-		return nil, ErrCreateApplication
-	}
-
-	return &application, nil
-}
-
-func (r *EventRepository) GetApplicationByUserAndEventID(ctx context.Context, params sqlc.GetApplicationByUserAndEventIDParams) (*sqlc.Application, error) {
-	application, err := r.db.Query.GetApplicationByUserAndEventID(ctx, params)
-
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			return nil, ErrApplicationNotFound
-		}
-
-		return nil, ErrGetApplication
-	}
-
-	return &application, nil
-}
-
-func (r *EventRepository) SubmitApplication(ctx context.Context, data any, userId, eventId uuid.UUID) error {
-	resumeUrl := "TODO"
-
-	jsonBytes, err := json.Marshal(data)
-
-	if err != nil {
-		return ErrInvalidApplicationData
-	}
-
-	err = r.db.Query.UpdateApplication(ctx, sqlc.UpdateApplicationParams{
-		StatusDoUpdate:      true,
-		Status:              sqlc.ApplicationStatusStarted,
-		ApplicationDoUpdate: true,
-		Application:         jsonBytes,
-		ResumeUrlDoUpdate:   true,
-		ResumeUrl:           &resumeUrl,
-		UserID:              userId,
-		EventID:             eventId,
-	})
-
-	if err != nil {
-		return ErrSubmitApplication
-	}
-
-	return nil
-}
-
-func (r *EventRepository) SaveApplication(ctx context.Context, data any, params sqlc.UpdateApplicationParams) error {
-	resumeUrl := "TODO"
-
-	jsonBytes, err := json.Marshal(data)
-
-	if err != nil {
-		return ErrInvalidApplicationData
-	}
-
-	err = r.db.Query.UpdateApplication(ctx, sqlc.UpdateApplicationParams{
-		StatusDoUpdate:      true,
-		Status:              sqlc.ApplicationStatusStarted,
-		ApplicationDoUpdate: true,
-		Application:         jsonBytes,
-		ResumeUrlDoUpdate:   true,
-		ResumeUrl:           &resumeUrl,
-		UserID:              params.UserID,
-		EventID:             params.EventID,
-	})
-
-	if err != nil {
-		return ErrSaveApplication
-	}
-
-	return nil
 }
