+++ apps/api/internal/services/events.go
@@ -180,47 +180,3 @@ func (s *EventService) AssignEventRole(
 
 	return nil
 }
-
-func (s *EventService) CreateApplication(ctx context.Context, params sqlc.CreateApplicationParams) (*sqlc.Application, error) {
-	application, err := s.eventRepo.CreateApplication(ctx, params)
-
-	if err != nil {
-		s.logger.Err(err).Msg(err.Error())
-		return nil, err
-	}
-
-	return application, nil
-}
-
-func (s *EventService) GetApplicationByUserAndEventID(ctx context.Context, params sqlc.GetApplicationByUserAndEventIDParams) (*sqlc.Application, error) {
-	application, err := s.eventRepo.GetApplicationByUserAndEventID(ctx, params)
-
-	if err != nil {
-		s.logger.Err(err).Msg(err.Error())
-		return nil, err
-	}
-
-	return application, nil
-}
-
-func (s *EventService) SubmitApplication(ctx context.Context, data any, userId uuid.UUID, eventId uuid.UUID) error {
-	err := s.eventRepo.SubmitApplication(ctx, data, userId, eventId)
-
-	if err != nil {
-		s.logger.Err(err).Msg(err.Error())
-		return err
-	}
-
-	return nil
-}
-
-func (s *EventService) SaveApplication(ctx context.Context, data any, params sqlc.UpdateApplicationParams) error {
-	err := s.eventRepo.SaveApplication(ctx, data, params)
-
-	if err != nil {
-		s.logger.Err(err).Msg(err.Error())
-		return err
-	}
-
-	return nil
-}
